breaking: only consider non-array objects to be valid "record"s
This does not match TypeScript quite as well, but it results in much clearer error messages, and generally enforces much closer to the behaviour most people would expect. The only thing this really impacts is people who were using the empty record to represent "any value except null or undefined". For this scenarios you might as well just allow null and undefined and use `any`.